Produce debuggable output when a kernel 'make' command fails
Previously the returncode of kernel builds and the output was
not considered and/or hidden. This made it hard to debug what
was going on and why a build failed.

We use select here for visibility of progress of the build.
Once we have more commands that need this, the code here
could be refactored into a function.