Allow building out of tree
This is just basic good practice, not to mix your build artifacts
with your source code.

However in this case it has a concrete benefit, it allows a
container to work properly. A container can carry around the basic
kernel source code staged in its image, and still allow building
to happen in a throw away empty volume.

Later this can grow into work to have a possible prebuilt tree in
an additional volume or directory, thus likely speeding up the
build by only building the diff. We don't implement that yet with
this patch.