Refactor patchwork patch retrieval
XMLRPC is a tricky protocol and it requires a lot of wrapper code
to avoid various problems. However, all of the information we need
for the patch itself is included in the mbox file that is downloaded
from Patchwork.

This patch removes the XMLRPC code and simply downloads the mbox file.
That file can then be parsed to get various information about the
patch itself.

This allows us to delete some code! Fixes #69.

Signed-off-by: Major Hayden <major@redhat.com>