Attach console log for relevant failures
So far, we sent out failure emails even if the problem occurred for the
previous kernel which we weren't testing. Avoid doing that. We also
attached the console log of only first job if it failed and ignored any
consequent failures, instead of attaching it for all other falied jobs.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>