Allow user to specify make job count
Users need the ability to increase or decrease the amount of CPU
cores used for kernel builds. This patch adds a `--jobcount` option
for `skt build` that allows a user to specify the amount of CPUs to
use. If a job count is not specified, `skt` will use the number of
CPUs installed in the system as a default.

Fixes: #91

Signed-off-by: Major Hayden <major@redhat.com>