Create separate state-related functions
Keeping configuration and state together in one file causes confusion
and makes troubleshooting challenging. It also mixes data that changes
often (state) and data that changes infrequently (configuration).

This patch adds new functions for destroying, reading, and saving
state to a YAML file. It also adds the pyyaml package to the
list of skt requirements.

Signed-off-by: Major Hayden <major@redhat.com>