Fix report message if no patches were applied
If the merge was tested and no patches were applied, the message still
said 'We applied the following patch...' and '... tests we ran on a
patchset you are involved with'. Both parts are misleading and wrong
since there are no patches.

Check if there are any patches listed first and choose appropriate
wording based on the data.

Fixes #110

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>