Update state_file api
As part of the split from storing state info in the config file to
a separate state file, the state_file api needs to be updated.

The main change is using the 'state' dict as the input instead of
the cfg dict.  The other change is tweaking the read() input to be
the path string instead of a dict.  This makes it more natural to
return a dict.

Another api change is the update() function.  The in-memory state data
needs to be updated, do this by returning a pointer to the new contents.
Updating by reference inside the function is to messy.

Most of the changes are mechanical: convert cfg -> state.

The tests were updated to reflect the new api.

Because this api is not being used, there is no risk of regression
with this patchset.