Use new state_file logic
Cut over to using a separate state_file.  Instead of saving to
the 'state' section in the config file, use a separate state file.

save_state(data) -> state = state_file.update(state, data)

Mechanical changes.

The only tricky part is the initial state_file setup.  Only read the config
file if the state file does not exist.

One odd thing about using cli args in state, is the 'func' arg.  That is a function
pointer and yaml.load is not a big fan of it.  So just delete it before storing
in the yaml file.