Remove leftover config state info
Now that the state info is moved into a separate file, remove
various logic that looked for 'state' info in the config file.

Most of the removal is in load_configs.  Now that it is not possible
to have 'state' info in the config file, remove the checks that
looked to see if certain fields existed or not.

Finally remove the original save_state function and the now unused
ast module.