Fix kernel config issues in reporter
The name of the attachment changed from 'config' to 'config.gz' as we
decided to compress the data, which wasn't reflected when working with
the attachment later. That said, printing gzip-ed data in StdioReporter
is fairly useless too, so remove that.

Add a comment to explain moving config attachment to the end of
attachment list, and iterate over the list of attachments instead of
checking only the first one, in case the order of adding attachments
changes.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>