test_runner: Add run tests
This is an initial attempt to test the 'run' cmd.

It even tests the 'wait' command, by assuming it passes on the first run
so we do not have to incur the wait delay.

This requires better test data.  Add a valid pass xml testcase to use.  This
allows the code to parse the data correctly.