Add new argument option --target
Using rhconfig_glob is specific to Red Hat configs as is not intuitive
with similar uses across other applications.

Instead use a more generic option '--target'.  This allows the user
to specify what they want and let skt map that into something useful.

Now instead of having Jenkins set the ARCH, skt now sets it correctly.

This makes using the cli easier as it relies less on outside settings.

The result of this patch is the removal of the short-lived rhconfig-glob
and replaced with a verify_target() function which maps the target to the
right kernel ARCH and config arch name.

Then cfgtype functions can take the config_arch and map them into whatever
cfgtype use case they need.  For rh-configs it is a direct mapping and
cleanly handles x86_64-debug and s390x-kdump correctly.

Tests are added to verify.