Use command line args for report
Writing JSON or Python dictionaries on the shell command line makes
it difficult to run reports. It is very easy to make mistakes in
quoting and it's difficult to check the arguments until you reach the
end of the process.

Add arguments for each of the reporters and restrict the reports which
can be chosen. Use importlib for instantiating the reporter class and
handle failures gracefully.

Allow command line arguments to take precedence over configuration files
for reporting.

Allow users to specify multiple email headers for messages.

This fix is related to #125.

Signed-off-by: Major Hayden <major@redhat.com>