WIP: Add stub code to test general workflow with dummy data
Testing the tools is hard when you need a lot of pieces to run.  Further
the amount of time it takes for those pieces to complete is long.

Short-circuit it, by quickly return the success/failure you want.

WIP because it is quick and dirty way of updating the state fail and
properly generating the expected report response.

Use:

skt --state --test 0 merge   # generates a successful merge in state data
skt --state --test 1 merge   # generates a failed merge in state data
skt --state report  # will report on the stub'd state data