Avoid creating multiple subject headers
Python's email module creates a new header with same key instead of
overwriting the existing one. While technically this is valid since RFCs
permit multiple headers with same keys, having multiple subjects is not
expected, and the sending site grabs the first (empty) subject and is
fine with it.

Check if the subject is provided and only create the header with a value
we want, not empty strings. This also means we can remove the default
value since we need to do a manual check for it anyways.

Fixes #220

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>