Add overall status to subject line of email
When reading through 10s of status emails, it is hard to detect quickly
which jobs failed.  Those are jobs I want to prioritize to verify the
pipeline is working correctly.

Stick the status in the subject line as [PASS] or [FAIL].  This requires
stripping a prepending Re: and re-adding it.