reporter: Allow reporting without a kernel .config
Remove the `config` key from merge data since it is only used for
attaching the kernel config file to the report email. Rename
`get_kernel_config()` to `attach_kernel_config()` to better
describe what it does.

If the kernel config exists, attach it to the report email and
provide lines for the report. If it does not exist (for example:
merge is complete but compile has not started), skip the attach
process and don't return any lines for the report.

Fixes #223.

Signed-off-by: Major Hayden <major@redhat.com>