Add LTP lite reports on failures
LTP lite was added to tests that are being run, but the report wasn't
updated to include links to logs. Since console logs aren't very useful
with LTP most of the time, these logs are important. Because we skip
reporting of failures we had no console logs for (residue of only
boot-testing), we also reported a failure but without any indication of
what actually failed.

The change boils down to addition of BeakerRunner method to retrieve LTP
logs for each recipe in the jobs and adding them to reports.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>