reporter: Skip run results for merge failure
When a multireport has a merge failure, the test run results
are still printed in the report along with:

    Patch application failure!

The report already has a paragraph explaining that the merge failed
and there's no need to print empty test runs below that paragraph.

Show the test run report headers only when the merge succeeded.

Fixes #256.

Signed-off-by: Major Hayden <major@redhat.com>