Fix infinite resubmission loop in BeakerRunner
If the job faield, it gets resubmitted on the same host. If the machine
gets removed from the pool (or anything else happens, making it
unreachable) and the resubmitted job aborts, BeakerRunner clones the
recipe and tries to submit it again, this time not caring about which
host it's on. However, the hostRequires in the recipe is set to request
a specific host from previous resubmission and hence copied over, making
BeakerRunner stuck trying to resubmit the recipe on the nonexistent host
all over again.

Remove hostname from the hostRequires from recipe XML and only add it
if `samehost` resubmission rule is True, to avoid this bug.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>