Update KeyboardInterrupt catching
Since we moved to installable which uses `main()` as entry point, the
code for catching KeyboardInterrupt was unused and the custom exception
handling was ignored and never called. Move catching of the exception to
the `main()` itself, and change the return code to 130 which is returned
by bash on CTRL-C pressing as well, for consistency.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>