Make __getresults recognise aborted jobs
If all jobs get aborted, something is wrong with the infrastructure.

To recognise these errors from real failuers, we need to check the
(result, status) values from the job result XML, and they must match
'Warn' and 'Aborted' ('Warn' itself is a legitimate test result and
'Aborted' is present with kernel panics as well - but in that case the
result is 'Panic').

Add the 'status' into failure info and check it to return appropriate
return code so we know the testing process failed and needs to be
repeated.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>