Fix failure info in multireport summary
While we return an error at the end of the runner stage if all the jobs
abort, it's better to account for the fact that somebody wants to run
the reporting by hand with those results. In those cases, we shouldn't
be saying that the testing failed since the testing didn't actually
complete.

Moving the setting of test failure summary into the __getjobresults
method with the additional checks also makes sure we don't report test
failure eg. if the build failed (we want to report the earliest
failure), or if the testing stage didn't run (there is no retcode).

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>