Exchange LTP logs
Exchange the LTP install log for the summary of run tests, and switch
the order of the logs. This way, the first linked log contains the test
summary where the failed tests can be found, and the second one the
details of test runs (which can be easily searched for the names of the
failed tests from the first log).

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>