Remove lastsubmitted from skt.runner.BeakerRunner
The lastsubmitted field in sktm.runner.BeakerRunner seems to have
been added for convenience of not needing to pass jobid. However,
this kind of logic is mostly useful for human interfaces, in an
internal interface it just complicates things and creates
confusion. Remove lastsubmitted and require always specifying
jobid.

Fixed #281