Remove unused __getconsolelog method
The method is retrieving the console log URL for the first recipe in the
job which isn't that useful for us. We need to have the console log for
each of the recipes that ran, and we are already doing that in a
different method. It doesn't make sense to keep this method around.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>