Remove support for infourl / buildinfo
Only duplicates information from the statefile, which is being passed
to the reporter anyways, and only contains the list of successfully
applied patches, as opposed to the statefile which also tells us which
patch failed to apply so we can report it to the developer. So unless
the merge succeeded, it makes no sense to use it, and even then it
contains the same data we already have in the statefile.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>