BeakerRunner.dumpjunitresult() accept output filename
At the moment, BeakerRunner.dumpjunitresults() accepts a directory
to write the junit file to, and it just assumes that it can
generate a sufficiently distinct filename to not clash with
existing files. It cannot know that, since it's not actually
creating the directory and filling it with files. Instead, the
caller should decide how the output file should be named, and the
function should accept a full name of the file to write to.

Fixes #273