Don't abort if no logs are found
If the job aborts with kernel panic before LTP completes, there won't be
any LTP logs present. The situation resulted in exception when trying to
report the results and not sending out kernel panics. Instead, let's try
to format the log URLs only if there are any present and fix this bug.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>