Remove dumpjunitresults method
There is no distinguishing between 'normal' jobs and the ones that got
aborted (and which we ignore). All of these get saved and then Jenkins
marks the result as failed, even though the return code says success.
Ignoring the return value of the runner and assuming things went wrong
because of temporary infrastructure failure isn't good. Also, with the
junit wrapper, there is no need to dump separate job results.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>