Make ConsoleLog accept file path as well
Until now, only URLs to console logs were accepted. Since the plan is to
move to separate stage reporting (in which compressed console logs will
be created as files), it will be useful to be able to grab and parse
these files, instead of awkwardly passing the URLs to logs.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>