Publisher.py: Append a slash (/) to the end of the destination path
If the destination directory supplied to skt publish doesn't exist,
skt creates a file with the supplied name. Instead, it should abort
with an error about non-existent directory. The easiest way to do
that would probably be appending a slash (/) to the end of the
destination path before transfering file.

Fixes #327