Enhance merge patch by applying the wildcard, or a series of patches
This patch allows user to merge a series of patches, and to use the
command with wildcard to merge. Because developers are more likely
to merge a series of patches, and the name of each patch is very
long, merging a series of patches is convenient for this purpose.
By using 'nargs="*"', the patches are stored in a list under
multiple optional arguments. In the class AppendMergeArgument,
each patch_file in values is going to add to the namespace by order.
At previous the old patch only is applied under each optional
arguments one time.

For example:

    --patch PATCH1.patch --patch PATCH2.patch --patch PATCH3.patch

Now, it is enhanced by applied a series of patch:

    --patch PATCH1.patch PATCH2.patch PATCH2.patch

Or it probably is merged by using wildcard command:

    --patch *.patch