Don't catch exceptions when aborting patch application
If `git am --abort` goes wrong, something is seriously wrong and we
shouldn't continue as if it wasn't. Instead of catching the exception,
let skt abort when it happens.

Doing so allows to distinguish between a merge failure and error, which
means the (possibly serious) issue won't go unnoticed. The original
patch application failure is also logged before aborting so the issue
can be reproduced.

Only continue the normal flow of reporting the patch application failure
if the cleanup went fine - we have no idea why `git am --abort` went
wrong, and it could be caused by a previously messed up repository, in
which case the application error is not caused by the patch itself, and
the problem should *not* be reported to its author.

Fixes #229

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>