Remove cmd_cleanup
This patch removes cmd_cleanup because:

1. The --wipe option is pretty drastic, especially if the working
directory is a cached git repo or contains additional files, or
directories that the user forgot about; and hence is not used in
practice.

2. We are slowly migrating to input/output directory structure
and workdir usage will be deprecated. Since each stage can use
different directories, it will become impossible to clean up
without specifying each of them on the command line of the cleanup
command. At this point, the command won't be any easier than
manual removal of the residue.

3. Related to the above, the state will not need to be used
anymore either, and hence won't need to be removed in this command.

Fixes #342