Make tarball name conform to the new output naming system
Instead of using the data from state (which will be removed) or
timestamp, make the tarball name "build.kernel.tar.gz". The "build"
prefix makes it clear that the file is related to the build stage,
while the rest of the name is descriptive enough to be used by humans.

This allows to remove the function for adding the timestamp to the
filename.

Use a more descriptive name instead of "ttgz".

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>