Make the kernel config filename conform to the new naming system
If the state is used, a commit hash is part of the config filename. If
it is not, "None" is used. Instead, use a "build" prefix to show that
the file is related to the build command, and make the rest of the name
descriptive for humans - it is a text file containing the configuration
used to build the kernel (to eg. distinguish from skt's own
configuration).

Also rename the "tconfig" variable to more descriptive "config_name".

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>