publisher: add prefix option
Allow the user to specify a filename prefix to use when publishing
files. This can be used to distinguish between runs if all are being
published in the same location.

For `cp` and `scp` publishers, this can be integrated into the
`destination` path creation but for `sftp`, the remote filename argument
needs to be added to the call directly.

Also extend the test for the `Publisher.geturl()` method to verify
correct URL is returned for both cases - if the prefix is or is not set.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>