Create the workdir right after it's name is retrieved
The workdir was only created during kerneltree initialization. This
leads to unexpected problem when the access rights to nonexistent
directory are checked during config loading, resulting in assignment of
current directory to input/output ones instead of the workdir.

Since kerneltree should now use merge.source as it's wdir, keep the
creation in init there and only remove the confusing comment. Add the
creation of the workdir itself into config loader to fix this bug.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>