Link machine info instead of embedding it
The emails with embedded machine description in the body can get lengthy
and easy to get lost in. Instead, we should link the HW info - people
who are interested in anything besides the architecture can just click
on the log URLs to get secondary information.

In the future, link to the job in the dashboard should be present in the
report instead of this - especially useful for going upstream where
internal links don't work, and also because the Beaker logs are deleted
after a few weeks.

For now, let's start with making the emails more cohesive and easier to
parse and understand, and get people used to clicking on the links to
get to detailed information.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>