Link the kernel config instead of attaching it
Based on the feedback from developers, the report emails are too huge.
Most of the size is taken up with the kernel configuration. Instead of
copressing and attaching it, we can take the config URL and link it in
the report instead.

Note that this is only applicable for the old reporter which is
currently fully in use - the build stage doesn't have the published
config URL yet. This can be fixed later, when the dashboard will contain
all the logs and other information used, and the related job will be
simply linked in the report.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>