Exclude logs that aren't important
Some tasks report a lot of logs, or have a lot of subtasks. If only one
of the subtasks fails, it can be complicated to find the right log in
the bunch of randomly looking links.

Exclude harness and setup logs from the email report since they contain
no actual data about the test run, and any "install" subtask logs as
well. On top of this, only include the logs to the subtasks that
actually failed.

This reduces the amount of information in the email and makes it easier
to find the logs with the actual failure descriptions.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>