Use jinja2 templates for reporting
Altering the content of email reports is challenging because almost all
content changes require changing Python code. In addition, the report is
built in small pieces which must be run the correct order and this makes
it difficult to change the order of the items in the report. The report
also includes too much data and it should include only failures.

This patch includes the following high-level changes:

  * Methods return lists/dicts that are rendered by jinja2 templates
  * Retrieve and parse failure data from Beaker, not all data
  * Report only failures and report those failures first
  * Update tests to look for new strings in report output

Signed-off-by: Major Hayden <major@redhat.com>