Don't get runner results if tests didn't run
If someone runs `skt report` for a test where the runner never started,
skt will fail with an error when it attempts to get the runner results.
This might happen if a user did a merge and compile tests but was not
interested in running a hardware test.

Ensure that the runner is actually specified in the skt rc/state file
before attempting to retrieve runner results.

Signed-off-by: Major Hayden <major@redhat.com>