Make report template configurable
Users need a way to specify a certain template when rendering reports.
This allows for different types of reports to be sent and it allows for
certain sensitive information to be sent in some reports that should not
appear in other reports.

Rename the base report template and include it from a default report template.

Add a `--template` argument for `skt report` to allow a user to specify a
preferred template.

Signed-off-by: Major Hayden <major@redhat.com>