Fix test result reporting when kpkginstall fails
If the kpkginstall task failed (eg. a panic during first boot), the task
was added to failed_tasks list and forgotten. This was caused by the
`continue` line in the `if` clause.

The code responsible for retrieving the list of failed tests (if
kpkginstall passed) was wrongly formatted and thus ran all the time
instead of only when it should, which contributed to additional
confusion and may have been the reason for the aforementioned `continue`
line.

Fix both issues above and while at it, add the '(boot test)' string to
the kpkginstall task name in case the task failed as well (it was only
present when it passed, making it look like the installation itself
failed and not the subsequent boot).

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>