Fix test result reporting in the summary
Previously, only test results for failed tests were retrieved. However,
this was changed in 3870ef32 and subsequent commits and not reflected in
the template, resulting in reporting passed testing as failed.

Simulate `if any()` condition to see if any of the recipes in the job
failed, and work around jinja's variable scope; to correctly detect the
passed and failed jobs.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>