Raise an exception if the replacement value is not a string
It is expected that the placeholders in the Beaker XMLs are replaced by
the values grabbed from the state file. However, because of a bug in skt
or manual testing process, the state file might not contain them. Trying
to use `None` as the value results in a cryptic `expected a character
buffer object` error. We need to use a cleaner message to indicate
what's wrong.

OTOH, the values typically grabbed from the state file might not be
needed because the XML already contains the right strings. Therefore, we
shouldn't raise the exception if the values in the state file are not
present, but only if we try to actually replace a placeholder with the
nonexistent value.

Fixes: #388

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>