Don't report aborted tasks
The task should abort if there's and infrastructure failure or a kernel
panic happened before and the task didn't run. Since the result of this
task is not relevant to the developers, it shouldn't be included in the
report.

Add a Warn/Aborted condition and clean up the if/elif/else condition to
use `==` instead of `!=` so it's easier to understand and manage.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>