Find kpkginstall task by the fetch URL
Beaker allows specifying a random string as a task name if it's being
fetched from a repository. This means we won't always be able to find
the kpkginstall task (which we care about) by hardcoded 'name'
attribute.

Since the task is provided by the CKI team, the fetch URL should be
stable to use. For people interested in caching the repository on their
own servers, it's not suitable to use the full URL -- however, we can
use the task's 'real' name included at the end of this URL which stays
the same.

This also allows for custom kernel install tasks - the task path/name in
the fetch URL needs to include "kpkginstall" but otherwise people are
free to use their own task for installing the targeted kernel.

This will also make our own lives easier if we need to include a task
for installing kernel RPMs instead of tarballs in the future - we are
all set by naming it 'kpkginstall-rpm'.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>