Add git fetch URL to report
It's useful for people to know what are the tests we actually ran,
especially if they failed and they'd like to reproduce the results.

This was included in the reports before but got lost with the addition
of the jinja templates.

Since now we are retrieving multiple attributes applicable to both
passed and failed tasks, unify this extraction so we don't need to keep
the same code twice.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>