feature: add redis soaking stub
This is a partial implementation of test soaking ("onboarding waived tests").

We use redis inside the container to load information about tests that
are soaking. The soaking information inside redis is a json-encoded dict
like:

    data = {
        "soaking": {
            "soaking_since": "09-01-2019",
            "runcount": 0,
            "enabled": 1
        }
    }

This patch modifies skt in such a way, that:
* if a test succeeds, we don't change anything
* if a test fails, skt looks for the testname above and tries to get soaking
  information; if there is any "enabled" attribute in soaking information
  at all, the test will be hidden if it fails/aborts

* failing soaking tests will not show in the report

The behaviour describe above will be disabled if --nosoak is used. The same
will happen if redis isn't available.
In order to enable/disable tests, they need to be added/removed in kpet-db.

Test coverage is included.

TODO (follow-up patches needed)
* gather information about failing/succeeding soaking tests
  and updateprocess it
* report on soaking tests (e.g. another email)

V1:
* more of a RFC patch
V2:
* --nosoak -> --soak
  - soaking is still enabled by default, disable using --soak false
V3:
* add missing 'soak' in test_reporter setUp()
V4:
* frigging whitespace
Signed-off-by: Jakub Racek <jracek@redhat.com>

Signed-off-by: Jakub Racek <jracek@redhat.com>