Refactor tarball searching into its own method
Move the tarball searching logic into its own method, `find_tarball()`.
This allows for more log searching capabilities later, such as searching
for RPM files.

Also, adjust append_and_log2stdout to just write to stdout. This allows
us to read the buildlog instead of storing a large list of log lines
while the build runs.

Add a test for the situation where a kernel was built, it appears in the
log, but the file doesn't actually exist on the filesystem.

Signed-off-by: Major Hayden <major@redhat.com>