keep every waived testrun in redis
This patch modifies how and what data are stored in redis.

Original idea was to store runcount and startdate for waived tests
in redis and when the test failed/aborted, suspend updating stats and
have someone review the results, then re-enable.
This is a little fragile, so instead we'll store every waived test run
in redis and make conclusions based on what the data shows us.

Signed-off-by: Jakub Racek <jracek@redhat.com>