fixup two result check bugs
1) check for 'Skip' task result

Even before __handle_test_fail() was introduced, the original code would
happily report any state different than 'Pass' as a test_failure,
leading to aborting that job.

2) a typo 'Failed' -> 'Fail'

err_on_failing_tasks() is there to handle waiving. It should check for
'Fail' not 'Failed' result of tasks.
This doesn't impact the code, because we don't yet use XMLs with _WAIVED
attribute and the failure is caught by a follow-up check (the code
adjusted for bugfix 1).

This patch contains testcase for bugfix 1) that I've seen and reported.

Signed-off-by: Jakub Racek <jracek@redhat.com>