reporter: Put all tasks in the same list
Put all the tasks (both passed and failed) into the same list in the
Repoter, and have a task property to describe if it passed or failed.
Similarly, extract logs for all tasks, not just failed. Filter the task
by properties when processing and rendering the template.

This results in shorter, simpler code, and prepares it for the addition
of waived task output.