kernelbuilder: Add support for tarball append
Allow for the distribution of additional data with the kernel tarball;
data that wouldn't ordinarily have made it to the tarball without a
modification to the Makefile itself. Can be used to distribute
miscellaneous/autogenerated files.

Currently, only .tar and .tar.gz files are supported.

Add a helper function that can be used to locate files to be appended.

Signed-off-by: Čestmír Kalina <ckalina@redhat.com>