Only forget the taskspec after resubmission
If we remove the recipe set and job from watchlist first, we fail to
find them in get_recipeset_group. This means that all the jobs will be
without the orginal group, because we don't know what job to query to
find it.

Move taskspec deletion after (potential) resubmission to fix this.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>