further adjust ewd check
This patch adjusts ewd check done in order to detect a likely kernel
failure. The problem with current logic is that when an installation
aborts, all tasks may have ewd hit.

This patch changes detection in a following way: if ewd is hit on all
tasks (including boottest), then this isn't considered to be a kernel
failure.

Signed-off-by: Jakub Racek <jracek@redhat.com>