drop unused options
There are some options that were already removed, or that we never set
in pipeline. Other, like specifying --runner are to be removed soon,
because we're using only 1 type of runner and it's very unlikely that
we'll be adding more runners anytime soon.

--state should not be necessary either, because we want to always ensure
skt dumps its results in one form or other.

Also, skt isn't supposed to handle all pipeline stages (merging,
building, ...) anymore, so we're really trying to make it as simple as
possible and only allow passing options that we use.

Signed-off-by: Jakub Racek <jracek@redhat.com>