fix multirecipe handling
SKT currently decides its retcode based on the first recipe it checks.
If there are any failures, it may prefer SKT_FAILURE over SKT_ERROR or
vice-versa.
This patch attempts to fix that by getting all the individual retcodes
from recipes and deciding result by the presence of a priority return
code.
The log message is modified to include recipeid that hit the priority
check.

Signed-off-by: Jakub Racek <jracek@redhat.com>