don't discard recipeset when we ran out of retries
The logic in SKT watchloop is convoluted and needs to be rewritten
eventually. For now, let's fix this bug. Whenever we do a next
abort-retry, we also delete the taskspec from both watchlist and from
results. This makes sense when we allow more retries, but when we run
out, we must keep the last result to be detected as infrastructure
issue.
The code as-is used to work only because when we run out of
abort-retries, we cancel remaining jobs and detect a cancelled run.
However, if we have only 1 job running, this doesn't work.

Signed-off-by: Jakub Racek <jracek@redhat.com>