make sure state is saved on abrupt exit
In our pipeline, skt is being run with timeout utility to kill it after a
set timeout. This is used to timeout on infrastructure issues, but we
currently don't write out skt state to rc-file, so we don't get even a
partial report for that given arch.

This patch makes skt always call cleanup_handler() once on exit.
If the run isn't finished, this will result in SKT_ERROR retcode being
set and the rest of the jobs being cancelled.

The patch is structured in following way:
* move cleanup_handler() and signal_handler() to executable.py
* move code that saves state to cleanup_handler()
* add condition to check that cleanup_handler() was run only once
* adjust tests

No other functional changes intended.

Signed-off-by: Jakub Racek <jracek@redhat.com>