fix termination conditions
This patch fixes 4 subtle and related bugs.

We test when we've run out of attempts. However, because it's not the
only increment in the code, we can get into a situation where this
condition can no longer succeed. Let's fix that and end the watchloop
with a return.

We remove recipe_set_id from watchlist on 3 different places in total.
As a result, we may end-up trying to remove a recipe_set_id from watchlist,
that is not there anymore.
Quietly catch KeyError exception and keep going.

When we exit the watchloop, we have to process the results and evaluate
the overall testing status. Because there are multiple recipesets in a
job, we end-up missing at least 1 recipeset in a multi-recipeset job,
as we haven't saved results for it in the list. So instead, we always
save these results as we fetch them.

Finally, we fix some unit tests that have been quietly failing because
of an exception, the SKT's returncode being the same only because it was
evaluated as an infrastructure issue.

See
https://xci32.lab.eng.rdu2.redhat.com/cki-project/cki-pipeline/-/jobs/692695/

Signed-off-by: Jakub Racek <jracek@redhat.com>