don't log traceback of signal_handler
When we do our signal_handler() to do cleanup and sys.exit(SKT_ERROR) on
timeout, we end-up catching the SystemExit exception and logging
the traceback. Also, we don't really return SKT_ERROR, that's likely set
by pipeline.
This has no consequences for testing itself, but scripts
that look at output would think this traceback is an indication of far,
which isn't far from the truth. So let's fix this cosmetic bug, log this
and reraise the exception.

Signed-off-by: Jakub Racek <jracek@redhat.com>